irqchip: irq-crossbar: Add irq number 133, 139 and 140 to the reserved list
irq 133 (default mapped to NMI pin), 139 and 140 behaves same as irq 10.
Any device which gets mapped to these irqs does not seem to be able to
generate an interrupt. For now suspecting an hardware issue. Mark
these as reserved for now.

These are based on trial and error/testing done by Nishanth/Suman/Sricharan
and we would still need more details (which are being followed upon with the h/w
teams) on the behaviour before this or anything
similar is posted upstream. So the intent of this patch for now is to just
pull the fixes for the know issues into internal trees for j6eco bringup
and validation and j6 validation using a 3.15 Intergration tree.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>