mfd: tps65917: remove caching regmap
Currently we select REGCACHE_RBTREE with the hope to reduce I2C
operations. However, we ignore the registers that have status bits!

Here is what happens as a result with caching, the eventual register
result is flushed out to the register at points of change as needed,
however reads are cached as well.

For example:
LDO1_CTRL register contains: bit 4 containing current LDO status, and
bits 0 and 2 containing active and sleep mode settings.
regulator driver does the following:
	- write to bit 0 to enable the regulator and reset it to disable
	it
	- read bit 4 to check regulator enable status.

However regmap cache does not show the current state of the register,
instead it provides the modified value. so when we first check the
status, say the value was 0x11, we disable the regulator(write 0x10),
the status inside the PMIC becomes 0x00 (status is disabled), but
cache indicates 0x10.

Further, at later operations, this cached result is pushed to the
register-> that is again not correct.

To fix this behavior, judicious use of regcache_sync*,
regcache_cache_bypass should be used. This must be carefully used
depending on the type of operation involved. However, the benefit of
cached operation for a regulator that is hardly modified is, at the
best, dubious in value.

So, just dont do cached operations for this chip as it provides hardly
any value and just complicates logic.

Signed-off-by: Nishanth Menon <nm@ti.com>
Acked-by: Keerthy <j-keerthy@ti.com>