ARM: DRA7/AM43XX: fix header definition for omap44xx_restart
omap44xx_restart is defined as a static void inline when DRA7/AM437X is
defined alone, which implies that the restart function is no longer
functional even though it is built in. So, fix the definition of the
same.

Unfortunately, the number of SoCs starting to use omap44xx_restart is
starting to grow and the definition looks a little ugly.

Signed-off-by: Nishanth Menon <nm@ti.com>